Updated declared compatible Blender version to 5.2.0: - Fixed model and animation export failing with a JSON serialization error caused by mathutils types switching to single precision (float32) in Blender 5.0. Exported values are now converted to plain Python floats. - mathutils-based arrays are now explicitly converted to float64 when passed to numpy, so their types match the type annotations (NumpyTable) and the float64 behavior from Blender 4.5. - Switched tests and type checking to Python 3.13 (the version bundled with Blender 5.2).

+++ mcblend/operator_func/animation.py
@@ -263,9 +263,11 @@ class Pose:
                 # Scale
                 local_matrix = objprop.get_local_matrix(
                     objprop.parent, normalize=False)
-                scale = np.array(local_matrix.to_scale())[[0, 2, 1]]
+                scale = np.array(
+                    local_matrix.to_scale(), dtype=np.float64)[[0, 2, 1]]
                 # Location
-                location = np.array(local_matrix.to_translation())
+                location = np.array(
+                    local_matrix.to_translation(), dtype=np.float64)
                 location = location[[0, 2, 1]] * MINECRAFT_SCALE_FACTOR
                 # Rotation
                 rotation = objprop.get_mcrotation(objprop.parent)
